"fix: laplafit - change optimization routine"
This commit changes the main module of the optimization routine
for the Laplace Symmetric distribution. Basically, we start
by the Maximum Likelihood estimator, and check the intervals
around to see if there is no new minimum.

While this seems redundant for the symmetric Laplace, it is
apparently not for the Asymmetric Laplace. For now, we will just leave
it here and check with several samples.