"fix: alaplafit - change optimization routine"
This basically implements the same optimization routine done for the
Symmetric Laplace distribution.

While in both cases there is a maximum likelihood estimate for the
m parameter, apparently in the asymmetric case the minimum is not
obtained there. I am not sure why, so I am going to investigate
more and test several more samples.